feat(plan16): align code with post-Plan 15 spec revisions
Apply 5 work items to align existing code with updated design specs
(daemon-architecture v1.0-r9, server-client-protocols v1.0-r13,
interface-contract v1.0-r11):

- ADR 00015: Protocol header 16→20 bytes, VERSION 1→2, sequence u32→u64
- ADR 00062: Split ratio f32→u32 fixed-point (x10^4), pure integer
  arithmetic in split_tree, navigation, notification_builder
- ADR 00003: Merge AttachOrCreate (0x010C/0x010D) into AttachSession
  (0x0104/0x0105) with session_name, create_if_missing, action_taken
- ADR 00059: Add caps_lock/num_lock to KeyEvent.Modifiers in libitshell3,
  libitshell3-ime; preserve wire bits 4-5 in wire_decompose
- Over-engineering fix: eliminate duplicate KeyEvent/ImeResult types by
  importing from libitshell3-ime; narrow hid_keycode u16→u8 with wire
  boundary validation per spec

1254 tests passing across 4 modules. Coverage: 95.37% merged.

+++ modules/libitshell3-protocol/src/error.zig
@@ -64,7 +64,7 @@ pub const ErrorCode = enum(u32) {
 /// 0x00FF. Error frame payload sent when a protocol or application error occurs.
 pub const ErrorResponse = struct {
     error_code: u32,
-    ref_sequence: u32 = 0,
+    ref_sequence: u64 = 0,
     detail: []const u8 = "",
 };
